Split up the different parts of the VitalSource service into files
