+++ lms/services/vitalsource/__init__.py
@@ -1,1 +1,3 @@
-from lms.services.vitalsource.client import VitalSourceService, VSBookLocation, factory
+from lms.services.vitalsource.factory import service_factory
+from lms.services.vitalsource.model import VSBookLocation
+from lms.services.vitalsource.service import VitalSourceService
