+++ tests/unit/lms/services/vitalsource/client_test.py
@@ -1,65 +1,10 @@
-from unittest.mock import create_autospec, sentinel
+from unittest.mock import sentinel
 
 import pytest
-from h_matchers import Any
 
 from lms.services.exceptions import ExternalRequestError
-from lms.services.vitalsource import VitalSourceService, factory
-from lms.services.vitalsource.client import VitalSourceClient, VSBookLocation
+from lms.services.vitalsource.client import VitalSourceClient
 from tests import factories
-
-
-class TestVSBookLocation:
-    TEST_CASES = [
-        ("vitalsource://book/bookID/book-id/cfi//abc", "book-id", "/abc"),
-        ("vitalsource://book/bookID/book-id/cfi/abc", "book-id", "abc"),
-    ]
-
-    @pytest.mark.parametrize("document_url,book_id,cfi", TEST_CASES)
-    def test_from_document_url(self, document_url, book_id, cfi):
-        loc = VSBookLocation.from_document_url(document_url)
-
-        assert loc == Any.instance_of(VSBookLocation).with_attrs(
-            {"book_id": book_id, "cfi": cfi}
-        )
-
-    @pytest.mark.parametrize("document_url,book_id,cfi", TEST_CASES)
-    def test_document_url(self, document_url, book_id, cfi):
-        loc = VSBookLocation(book_id, cfi)
-
-        assert loc.document_url == document_url
-
-
-class TestVitalSourceService:
-    def test_get_launch_url(self, svc):
-        document_url = "vitalsource://book/bookID/book-id/cfi//abc"
-
-        assert (
-            svc.get_launch_url(document_url)
-            == "https://hypothesis.vitalsource.com/books/book-id/cfi//abc"
-        )
-
-    @pytest.mark.parametrize(
-        "proxy_method,args",
-        (
-            ("get_book_toc", [sentinel.book_id]),
-            ("get_book_info", [sentinel.book_id]),
-        ),
-    )
-    def test_proxied_methods(self, svc, client, proxy_method, args):
-        result = getattr(svc, proxy_method)(*args)
-
-        proxied_method = getattr(client, proxy_method)
-        proxied_method.assert_called_once_with(*args)
-        assert result == proxied_method.return_value
-
-    @pytest.fixture
-    def client(self):
-        return create_autospec(VitalSourceClient, instance=True, spec_set=True)
-
-    @pytest.fixture
-    def svc(self, client):
-        return VitalSourceService(client)
 
 
 class TestVitalSourceClient:
@@ -161,20 +106,3 @@ class TestVitalSourceClient:
         HTTPService = patch("lms.services.vitalsource.client.HTTPService")
 
         return HTTPService.return_value
-
-
-class TestFactory:
-    def test_it(self, pyramid_request, VitalSourceService, VitalSourceClient):
-        svc = factory(sentinel.context, pyramid_request)
-
-        VitalSourceClient.assert_called_once_with(api_key="test_vs_api_key")
-        VitalSourceService.assert_called_once_with(VitalSourceClient.return_value)
-        assert svc == VitalSourceService.return_value
-
-    @pytest.fixture
-    def VitalSourceService(self, patch):
-        return patch("lms.services.vitalsource.client.VitalSourceService")
-
-    @pytest.fixture
-    def VitalSourceClient(self, patch):
-        return patch("lms.services.vitalsource.client.VitalSourceClient")
